assign entity options
