+++ lib/convertor/jdl_convertor.js
@@ -4,6 +4,8 @@ const fs = require('fs'),
     _ = require('lodash'),
     buildException = require('../exceptions/exception_factory').buildException,
     exceptions = require('../exceptions/exception_factory').exceptions,
+    BinaryOptions = require('../core/jhipster/binary_options'),
+    formatComment = require('../utils/format_utils').formatComment,
     JDLParser = require('../dsl/jdl_parser');
 
 module.exports = {
@@ -37,15 +39,14 @@ function convertToJson(intermediateData, databaseType, force) {
   /*var scheduledClasses = jdlObject.classes;
   if (parsedData.userClassId) {
     scheduledClasses = filterScheduledClasses(parsedData.userClassId, scheduledClasses);
-  }
+  }*/
 
-  var creator = new EntitiesCreator(parsedData, parser.databaseTypes);
-
-  creator.createEntities();
-  if (!force) {
+  createEntities();
+  /*if (!force) {
     scheduledClasses = creator.filterOutUnchangedEntities(scheduledClasses);
-  }
-  return creator.getEntitiesMap(scheduledClasses);*/
+  }*/
+
+  return entities;
 }
 
 
@@ -59,7 +60,7 @@ function createEntities() {
      * created JHipster User entity and none of its fields and ownerside
      * relationships will be considered.
      */
-    if (parsedData[entityName].name.toLowerCase() === USER.toLowerCase()) {
+    if (parsedData.entities[entityName].name.toLowerCase() === USER.toLowerCase()) {
       console.warn(
               "Warning:  An Entity called 'User' was defined: 'User' is an" +
               ' entity created by default by JHipster. All relationships toward' +
@@ -88,24 +89,37 @@ function initializeEntities() {
       relationships: [],
       fields: [],
       changelogDate: getChangelogDate(classId, index),
-      /*dto: getDTO(classId),
-      pagination: parsedData[classId].pagination,
-      service: this.parsedData.getClass(classId).service,
-      microserviceName: this.parsedData.getClass(classId).microserviceName,
-      searchEngine: this.parsedData.getClass(classId).searchEngine,
-      javadoc: formatComment(this.parsedData.getClass(classId).comment),
-      entityTableName: _.snakeCase(this.parsedData.getClass(classId).tableName)*/
+      dto: getOption(classId, BinaryOptions.BINARY_OPTIONS.DTO),
+      pagination: getOption(classId, BinaryOptions.BINARY_OPTIONS.PAGINATION),
+      service: getOption(classId, BinaryOptions.BINARY_OPTIONS.SERVICE),
+      microserviceName: getOption(classId, BinaryOptions.BINARY_OPTIONS.MICROSERVICE),
+      searchEngine: getOption(classId, BinaryOptions.BINARY_OPTIONS.SEARCH_ENGINE),
+      angularJSSuffix: getOption(classId, BinaryOptions.BINARY_OPTIONS.ANGULAR_SUFFIX),
+      javadoc: formatComment(parsedData.entities[classId].comment),
+      entityTableName: _.snakeCase(parsedData.entities[classId].tableName)
     };
 
     entities[classId] = initializedEntity;
   });
 };
 
-/*function getDTO(classId) {
+/**
+ * get the specified option for an entity if any.
+ * @param{String} classId  the id of the class to check for.
+ * @param{String} optionType the option type to fetch.
+ */
+function getOption(classId, optionType) {
+  let val;
   parsedData.options.forEach(function (option) {
-    if(BinaryOptions.BINARY_OPTIONS.DTO === option.name)
+    if(optionType === option.name) {
+      if (option.entityNames.has(classId) || (option.entityNames.has('*') && !option.excludedNames.has(classId))) {
+        val = option.value;
+      }
+    }
   });
-}*/
+  return val;
+}
+
 /**
  * If the entity already have a json file we get the changelogDate in it
  * else we create the changelogDate with liquibase date format
@@ -314,7 +328,7 @@ function setRelationshipOfEntity(classId) {
 function readJSON(classesToRead) {
   var entitiesRead = {};
   classesToRead.forEach(function (classToRead) {
-    var file = '.jhipster/' + parsedData[classToRead].name + '.json';
+    var file = '.jhipster/' + parsedData.entities[classToRead].name + '.json';
 
     if (onDiskEntities[classToRead]) {
       entitiesRead[classToRead] = onDiskEntities[classToRead];
