+++ lib/utils/objects/set.js
@@ -8,7 +8,7 @@ const buildException = require('../../exceptions/exception_factory').buildExcept
  */
 class Set {
   constructor(array) {
-    this.container = (!array) ? {} : convertToMap(array);
+    this.container = array ? convertToMap(array) : {};
   }
 
   has(element) {
